+++ database/migrations/2016_06_22_153608_add_config_option_for_ad_group_hierarchy_matching.php
@@ -1,0 +1,41 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddConfigOptionForAdGroupHierarchyMatching extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        if (!Schema::hasColumn('ldap_config', 'map_group_hierarchy')) {
+            Schema::table(
+                'ldap_config',
+                function (Blueprint $t){
+                    $t->boolean('map_group_hierarchy')->after('map_group_to_role')->default(0);
+                }
+            );
+        }
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        if (Schema::hasColumn('ldap_config', 'map_group_hierarchy')) {
+            Schema::table(
+                'ldap_config',
+                function (Blueprint $t){
+                    $t->dropColumn('map_group_hierarchy');
+                }
+            );
+        }
+    }
+}