+++ CHANGELOG.md
@@ -3,11 +3,17 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Added
 
 ### Changed
 
 ### Fixed
+
+## [0.3.1] - 2016-07-08
+### Added
+- DF-752 Added support for AD group hierarchy when mapping DreamFactory role to AD group.
+
+### Changed
+- General cleanup from declaration changes in df-core.
 
 ## [0.3.0] - 2016-05-27
 ### Changed
@@ -50,7 +56,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2015-10-24
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.3.1...HEAD
+[0.3.1]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.2.2...0.3.0
 [0.2.2]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/dreamfactorysoftware/df-adldap/compare/0.2.0...0.2.1